requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==0.27.0
python-dotenv==1.0.1